refactore email field components
