+++ src/components/emails/emailSignUp.jsx
@@ -1,6 +1,6 @@
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faCircleCheck } from '@fortawesome/free-solid-svg-icons'
-import signup from '../assets/images/signup.png'
+import signup from '../../assets/images/signup.png'
 import EmailField from './emailField'
 
 function EmailSignUp() {
