+++ src/components/emails/index.js
@@ -1,0 +1,4 @@
+import EmailField from './emailField'
+import EmailSignUp from './emailSignUp'
+
+export { EmailField, EmailSignUp }
